+++ src/main/java/io/github/astrapi69/swing/menu/model/MenuType.java
@@ -1,0 +1,9 @@
+package io.github.astrapi69.swing.menu.model;
+
+public enum MenuType {
+    SYSTEM_TRAY,
+    MENU_BAR,
+    MENU,
+    MENU_ITEM,
+    POPUP
+}